+++ packages/plugin-packages/mashroom-portal-default-theme/src/backend/theme-params.ts
@@ -1,6 +1,8 @@
 
 type ThemeParams = {
     spaMode: boolean;
+    darkMode: true | false | 'auto';
+    styleFile: string | null;
     showPortalAppHeaders: boolean;
     showEnvAndVersions: boolean;
     mashroomVersion: string;
@@ -8,6 +10,8 @@ type ThemeParams = {
 
 let params: ThemeParams = {
     spaMode: true,
+    darkMode: false,
+    styleFile: null,
     showPortalAppHeaders: true,
     showEnvAndVersions: false,
     mashroomVersion: '',
@@ -16,6 +20,12 @@ let params: ThemeParams = {
 export default {
     get spaMode() {
         return params.spaMode;
+    },
+    get darkMode() {
+        return params.darkMode;
+    },
+    get styleFile() {
+        return params.styleFile;
     },
     get showPortalAppHeaders() {
         return params.showPortalAppHeaders;